+++ pkg/ptrace/ptrace_linux_arm64.go
@@ -21,34 +21,36 @@ import (
 	"syscall"
 
 	"github.com/pkg/errors"
+	"golang.org/x/sys/unix"
 )
 
 var endian = binary.LittleEndian
 
-const syscallInstrSize = 2
+const syscallInstrSize = 4
 
-const nrProcessVMReadv = 310
-const nrProcessVMWritev = 311
+const nrProcessVMReadv = 270
+const nrProcessVMWritev = 271
+
+// see kernel source /include/uapi/linux/elf.h
+const nrPRStatus = 1
 
 func getIp(regs *syscall.PtraceRegs) uintptr {
-	return uintptr(regs.Rip)
+	return uintptr(regs.Pc)
 }
 
 func getRegs(pid int, regsout *syscall.PtraceRegs) error {
-	err := syscall.PtraceGetRegs(pid, regsout)
+	err := unix.PtraceGetRegSetArm64(pid, nrPRStatus, (*unix.PtraceRegsArm64)(regsout))
 	if err != nil {
 		return errors.Wrapf(err, "get registers of process %d", pid)
 	}
-
 	return nil
 }
 
 func setRegs(pid int, regs *syscall.PtraceRegs) error {
-	err := syscall.PtraceSetRegs(pid, regs)
+	err := unix.PtraceSetRegSetArm64(pid, nrPRStatus, (*unix.PtraceRegsArm64)(regs))
 	if err != nil {
 		return errors.Wrapf(err, "set registers of process %d", pid)
 	}
-
 	return nil
 }
 
@@ -67,26 +69,15 @@ func (p *TracedProgram) Syscall(number uint64, args ...uint64) (uint64, error) {
 		return 0, err
 	}
 	// set the registers according to the syscall convention. Learn more about
-	// it in `man 2 syscall`. In x86_64 the syscall nr is stored in rax
-	// register, and the arguments are stored in rdi, rsi, rdx, r10, r8, r9 in
-	// order
-	regs.Rax = number
+	// it in `man 2 syscall`. In aarch64 the syscall nr is stored in w8, and the
+	// arguments are stored in x0, x1, x2, x3, x4, x5 in order
+	regs.Regs[8] = number
 	for index, arg := range args {
 		// All these registers are hard coded for x86 platform
-		if index == 0 {
-			regs.Rdi = arg
-		} else if index == 1 {
-			regs.Rsi = arg
-		} else if index == 2 {
-			regs.Rdx = arg
-		} else if index == 3 {
-			regs.R10 = arg
-		} else if index == 4 {
-			regs.R8 = arg
-		} else if index == 5 {
-			regs.R9 = arg
-		} else {
+		if index > 6 {
 			return 0, errors.New("too many arguments for a syscall")
+		} else {
+			regs.Regs[index] = arg
 		}
 	}
 	err = setRegs(p.pid, &regs)
@@ -97,9 +88,9 @@ func (p *TracedProgram) Syscall(number uint64, args ...uint64) (uint64, error) {
 	instruction := make([]byte, syscallInstrSize)
 	ip := getIp(p.backupRegs)
 
-	// set the current instruction (the ip register points to) to the `syscall`
-	// instruction. In x86_64, the `syscall` instruction is 0x050f.
-	binary.LittleEndian.PutUint16(instruction, 0x050f)
+	// most aarch64 devices are little endian
+	// 0xd4000001 is `svc #0` to call the system call
+	endian.PutUint32(instruction, 0xd4000001)
 	_, err = syscall.PtracePokeData(p.pid, ip, instruction)
 	if err != nil {
 		return 0, errors.Wrapf(err, "writing data %v to %x", instruction, ip)
@@ -111,27 +102,26 @@ func (p *TracedProgram) Syscall(number uint64, args ...uint64) (uint64, error) {
 		return 0, err
 	}
 
-	// read registers, the return value of syscall is stored inside rax register
+	// read registers, the return value of syscall is stored inside x0 register
 	err = getRegs(p.pid, &regs)
 	if err != nil {
 		return 0, err
 	}
 
-	// restore the state saved at beginning.
-	return regs.Rax, p.Restore()
+	return regs.Regs[0], p.Restore()
 }
 
 // JumpToFakeFunc writes jmp instruction to jump to fake function
 func (p *TracedProgram) JumpToFakeFunc(originAddr uint64, targetAddr uint64) error {
 	instructions := make([]byte, 16)
 
-	// mov rax, targetAddr;
-	// jmp rax ;
-	instructions[0] = 0x48
-	instructions[1] = 0xb8
-	endian.PutUint64(instructions[2:10], targetAddr)
-	instructions[10] = 0xff
-	instructions[11] = 0xe0
+	// LDR x9, #8
+	// BR x9
+	// targetAddr
+	endian.PutUint32(instructions[0:], 0x58000049)
+	endian.PutUint32(instructions[4:], 0xD61F0120)
+
+	endian.PutUint64(instructions[8:], targetAddr)
 
 	return p.PtraceWriteSlice(originAddr, instructions)
 }